Move pyproject to project root
Split cuda file generation
Add custom build backend
Update installation document
